smart patient medicine tracker - Copy/requirements.txt
streamlit
opencv-python
face-recognition
pandas
openpyxl
qrcode
pillow
numpy
matplotlib
seaborn
